Update testing approach
